+++ internal/provider/provider.go
@@ -214,6 +214,7 @@ func (p *DependencyTrackProvider) Resources(ctx context.Context) []func() resour
 		NewPolicyResource,
 		NewACLMappingResource,
 		NewTeamAPIKeyResource,
+		NewUserTeamMembershipResource,
 	}
 }